création de navbar et partage de state
